+++ template/app/helpers.wasp.ts
@@ -1,0 +1,11 @@
+import { type Part } from "@wasp.sh/spec";
+
+export function group<SomePart extends Part>(
+  sharedState: Partial<SomePart>,
+  parts: SomePart[],
+) {
+  return parts.map((part) => ({
+    ...part,
+    ...sharedState,
+  }));
+}
